Broaden intended readership & utility of this doc
This doc is a very useful tool for guiding *anyone* reading a spec to focus their attention in the right places and to understand what questions to ask. It helps gets everybody on the same page—editors, WG members, and 3rd-party reader/reviewers—so it’d be helpful to make the abstract and intro more welcoming to readers, and more 1inclusive.