Revert "Merge pull request #10 from jasonanovak/week-02-introductions-and-threat-models"
Merging this PR resulted in my inital PR to the w3ctag version of the document
getting updated with the additional commits, when, actually, we want them to be
separate.

So, going to revert this PR merge, and then make a new PR of the week-2 branch to
the master branch of the w3ctag document getting updated with the additional
commits, when, actually, we want them to be separate.

So, going to revert this PR merge, and then make a new PR of the week-2 branch to
the master branch of the w3ctag documentt